Fix builtwithwicket carroussel by downloading images locally to comply with CSP of ASF
The Content Security Policy prohibits downloading resources from
external sites due to the GDPR. This prevented the Built With Wicket
carroussel to load images. This is fixed by downloading all images to
the site and serve them from the ASF servers together with the other
site resources.

Updating the site with content from tumblr should be as simple as
running the script `_tumblr/run_all.sh`. This will perform the following
tasks:

1. Fetch all posts from Tumblr
2. Download all images that have not yet been cached
3. Generate a tumblr.json file in the root with the latest 20 posts

This is tested running in a devcontainer in VSCode and should be
portable across all machines.